Added city generation.
Choosing one of the random city options on the main menu generates a
random location. The city generation works as intended for the most
part; just need to fix some starting locations and missing wet chasms.
Also added yellow text to world map for currently occupied province, and
added location name to status pop-up and automap. Also expanded voxel
size from 8 bits to 16 bits to allow for more than 256 voxel data in the
voxel grid.

+++ OpenTESArena/src/Interface/AutomapPanel.cpp
@@ -131,8 +131,8 @@ AutomapPanel::AutomapPanel(Game &game, const Double2 &playerPosition,
 
 		auto getVoxelData = [&voxelGrid](int x, int y, int z) -> const VoxelData&
 		{
-			const uint8_t *voxels = voxelGrid.getVoxels();
-			const uint8_t id = voxels[x + (y * voxelGrid.getWidth()) +
+			const uint16_t *voxels = voxelGrid.getVoxels();
+			const uint16_t id = voxels[x + (y * voxelGrid.getWidth()) +
 				(z * voxelGrid.getWidth() * voxelGrid.getHeight())];
 			return voxelGrid.getVoxelData(id);
 		};
